requirements.txt
asgiref==3.7.2
cloudinary==1.36.0
crispy-bootstrap4==2023.1
dj-database-url==0.5.0
dj3-cloudinary-storage==0.0.6
Django==4.2.6
django-allauth==0.58.0
django-crispy-forms==2.1
django-restframework==0.0.1
django-summernote==0.8.20.0
gunicorn==21.2.0
oauthlib==3.2.2
Pillow==10.1.0
psycopg2==2.9.9
PyJWT==2.8.0
python-magic==0.4.27
python3-openid==3.2.0
pytz==2023.3.post1
requests-oauthlib==1.3.1
sqlparse==0.4.4
urllib3==1.26.15
